feat(feedback): implement feedback form for each post (#37)
This PR implements feedback for posts. This PR utilizes Netlify Functions, which are implemented in Go (as I want to learn more about Go). I know it's more complex compared to JS/TS, but still, I want to try something new.

The database is Fauna. Again, something new. I want to try Supabase but maybe for another project.

## Feature

The feedback form is using a modal that is accessible by clicking the "Send anonymous feedback" at the bottom left of each post. On top of that, this PR also adds the "Edit this post" link, which if clicked, the browser will open a new tab that will edit the post's Markdown file.

The slug is automatically "attached" to the form. A validation (manually added) exists, which will protect the textarea from being submitted if it is an empty string.

The Fauna part is a bit tricky. Since I was new to Go, I needed to learn a lot of stuff (see https://github.com/imballinst/learn-golang), and especially FQL took quite a bit of my time. It wasn't as easy as GraphQL, I think. Nevertheless, I got it working for the GET (List) and POST (Create) comments.

At the moment, the GET endpoint isn't used yet, but probably it'll be used later when I want to create a dashboard for curating the feedback... one day™.

## Future works

Need to monitor the stuff and see if some kind of protection is required, e.g. spam. Spam is not OK in this case because I'm using Netlify functions which has 125k request limit.

Probably it's also a good idea to limit the content length so that the request size isn't too big.